Add more clang-tidy cleanup

+++ impl/jamtemplate/sdl/shape.cpp
@@ -31,12 +31,14 @@ jt::Vector2 Shape::getPosition() const { return m_position; }
 jt::Rect Shape::getGlobalBounds() const
 {
     return jt::Rect { m_position.x() + getOffset().x(), m_position.y() + getOffset().y(),
-        m_sourceRect.width() * m_scale.x(), m_sourceRect.height() * m_scale.y() };
+        static_cast<float>(m_sourceRect.width()) * m_scale.x(),
+        static_cast<float>(m_sourceRect.height()) * m_scale.y() };
 }
 jt::Rect Shape::getLocalBounds() const
 {
-    return jt::Rect { m_position.x(), m_position.y(), m_sourceRect.width() * m_scale.x(),
-        m_sourceRect.height() * m_scale.y() };
+    return jt::Rect { m_position.x(), m_position.y(),
+        static_cast<float>(m_sourceRect.width()) * m_scale.x(),
+        static_cast<float>(m_sourceRect.height()) * m_scale.y() };
 }
 
 void Shape::setScale(jt::Vector2 const& scale) { m_scale = scale; }
@@ -87,8 +89,8 @@ SDL_Rect Shape::getDestRect(jt::Vector2 const& positionOffset) const
     auto const pos = m_position + getShakeOffset() + getOffset() + getCamOffset() + positionOffset
         + m_offsetFromOrigin;
     SDL_Rect const destRect { static_cast<int>(pos.x()), static_cast<int>(pos.y()),
-        static_cast<int>(m_sourceRect.width() * fabs(m_scale.x())),
-        static_cast<int>(m_sourceRect.height() * fabs(m_scale.y())) };
+        static_cast<int>(static_cast<float>(m_sourceRect.width()) * fabs(m_scale.x())),
+        static_cast<int>(static_cast<float>(m_sourceRect.height()) * fabs(m_scale.y())) };
     return destRect;
 }
